Validate scope name components.
Previously we had a scope (gopkg.in) containing a dot. However dots are
used to nest scopes, so this was potentially ambiguous and a bug waiting
to happen (e.g., this only worked because we created an implicit 'gopkg'
parent scope behind the scenes, which isn't what was intended).

Now we forbid dots, and for uniformity we also require lower-case letters
and dashes instead of underscores.

Invalid scope names are still allowed, but deprecated. In 1.2 they will
become errors.

This change deprecates gopkg.in and, while we're at it, a couple of other
go-related scopes.

Unfortunately we can't do this check in one central place, because by the
time that would be possible we already have composite scope names (foo.bar)
which would fail the check. We have to check the individual components before
they get munged together, so it's a little less elegant than I would like.
Not terrible though.

Testing Done:
CI passes: http://jenkins.pantsbuild.org/job/pantsbuild/job/pants/branch/PR-3448/

Reviewed at https://rbcommons.com/s/twitter/r/3893/