New Google Analytics tracking code for www.pantsbuild.org.
This is tied to the Google Analytics account associated with
pantsbuild@gmail.com.

The old code was owned by someone at Twitter, and we don't know
who.  And in any case we want pantsbuild@gmail.com to own this
going forward, and we never looked at the old analytics (as we
had no access to the account), so creating a new tracking code
seems like the best way to go.

Testing Done:
N/A

I'll look at the analytics in a few days and make sure the data is flowing correctly.

Reviewed at https://rbcommons.com/s/twitter/r/3917/