Make `JvmAppAdaptor` compatible with bare `bundle()` form.
Given a BUILD file like:

```
jvm_app(
  name='bundle',
  basename='xyz',
  binary='x/y/z:bin',
  bundles=[
    bundle()
  ]
)
```

running `./pants --enable-v2-engine list <target>` currently throws an `AttributeError: JvmAppAdaptor(address=x/y:bundle) does not have attribute 'field_adaptors'`. this was an in-property exception (itself causing an `AttributeError` for failed resolution of the property) which actually turned out to be a second `AttributeError` for an invalid access of an undefined `bundle.fileset` attribute.

this review adds handling of this case with logging - as well as blanket exception logging for the `field_adaptors` properties to help reveal otherwise hidden exceptions in the future.

Testing Done:
reproduced locally in our monorepo against one of ~250 targets of this form. verified the exception logging prior to the change and the warning logging after the change.

https://travis-ci.org/pantsbuild/pants/builds/134654915

Bugs closed: 3540

Reviewed at https://rbcommons.com/s/twitter/r/3965/