Support for transitioning to making all backends opt-in.
Replaces the hard-coded list of backends with an already-deprecated
option.  A deprecation warning will issue if that option's values
aren't repeated in --backend-packages.

This allows repos to opt out of currently hard-coded backends,
by removing them from --default-backend-packages.  The deprecation
messages encourage repos to explicitly list the backends they
want in --backend-packages, in advance of switching to an entirely
opt-in system.

In short: this change allows repos to opt out, a future change
will require repos to opt in.