[engine] Add/use intrinsic for fingerprinting Files
This change allows codepaths that only need a digest of a file path to avoid holding the content of the file in memory. It's a stopgap improvement for memory usage until we began to incorporate the snapshot support that Nick is currently working on.

- Add support for requesting a FileDigest rather than FileContent.
- Use FileDigests to compute the EagerFilesetWithSpec that is used for BuildGraph integration.
- Improve trace output slightly (not enough to call https://github.com/pantsbuild/pants/issues/3535 complete)

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/135284005
http://jenkins.pantsbuild.org/job/pantsbuild/job/pants/branch/PR-3557/

Bugs closed: 3557

Reviewed at https://rbcommons.com/s/twitter/r/3974/