Delete the spindle-plugin from contrib.
We now publish this out of our monorepo, and the contrib
module has lagged behind. It is unlikely to get any further
updates from us and since it isn't providing much value
it should be deleted.

It used to be useful as an example of a contrib module
but now that there are many plugins that no longer
is unique.