Delete the spindle-plugin from contrib.
We now publish this out of Fsq.io, and the Pants contrib
module code has lagged behind. It is unlikely to get any further
updates from Foursquare and since it isn't providing much value
elsewhere we think it should be deleted.

It used to be useful as an example of a contrib module
but now that there are many plugins that no longer
is unique.

It is unclear if we have any process around halting updates
or deleting plugins. But a bunch plugins are only being
published because we hardcode the pantsbuild.pants requirement,
the plugins themselves are generally pretty stable.

Testing Done:
Passed Travis. (Jenkins failed b/c the node download failed, unrelated): https://travis-ci.org/pantsbuild/pants/builds/136236969
Ran a dryrun of the release.sh.

Bugs closed: 3571

Reviewed at https://rbcommons.com/s/twitter/r/3990/