A `remote_sources` target as a better mechanism for from_target.
Stu suggested in https://rbcommons.com/s/twitter/r/3830/ that we
overhaul the deferred sources system to generate synthetic targets
instead. This should eliminate a lot of special casing we've been
having to do around deferred sources, since we can use the same
logic that was already handling synthetic targets created from
generated code.

This change adds the `remote_sources` target as a replacement for
the `from_target` object, however it does not remote the existing
functionality of `from_target`. We can remove `from_target` in a
later (breaking) change.

In the mean time, I added an option to `deferred_sources_mapper`
to throw an exception if `from_target` is used, which will allow
individual users to eliminate their own usage of `from_target` in
favor of the more robust `remote_sources`.