When source fields are strings, not collections, raise an error; Test deferred sources addresses error
Currently if a string is passed to sources in a target definition, it will be treated as a list of single character filenames, which blows up in a confusing way later. This makes passing a string an error.

While I was in here, I added unit tests covering the too many addresses case for deferred sources and uncovered a bug where if the address kwarg isn't passed, instead of raising WrongNumberOfAddresses, it would instead hit an AttributeError on None.

Testing Done:
Wrote a few regression tests and made them pass.

Jenkins and travis passed

- http://jenkins.pantsbuild.org/job/pantsbuild/job/pants/branch/PR-3550/1/
- https://travis-ci.org/pantsbuild/pants/builds/135065849

Bugs closed: 3507, 3550

Reviewed at https://rbcommons.com/s/twitter/r/3970/

closes #3507, closes #3550