Provide a better error message when target `dependencies` is passed a string, not a list
Previously, a target of the form:

```
target(name='foo', dependencies=':bar')
```

was parsed as the dependencies ':', 'b', 'a', 'r'.  Now we check to see that dependencies
is a list, set, or tuple.  Pants exited cryptically with the error:

```
Exception caught: (<type 'exceptions.ValueError'>)
  ...
  File "/Users/zundel/Src/Pants/src/python/pants/build_graph/address.py", line 84, in parse_spec
    check_target_name(target_name)
  File "/Users/zundel/Src/Pants/src/python/pants/build_graph/address.py", line 71, in check_target_name
    raise ValueError('Spec {spec} has no name part'.format(spec=spec))

Exception message: Spec : has no name part
```

With no path to the BUILD file or indication of which target was wrong.

Also, catch spec parsing error and convert to AddressLookupError to at least get a path to error if something
like this slips through again.