Implicitly provide source products.
The ide project gen tasks require the 'java', 'scala' products.
The only reason this works is that the scrooge task happens to
provide those products.  But if we try to deregister the codegen
backend in some repo that doesn't need it, we get a MissingProductError
from the round engine, and a quite unintuitive error message.

This change makes every source root implicitly provide a product
with the same name as that root's language(s).  This makes logical
sense, and also solves the problem above.

While testing this, I noticed an error in how we iterate over all
source roots: We assumed that all fixed source roots come from
the appropriate options. But in fact they might also be added
programmatically.  So this change also fixes that, and adds
appropriate testing.

Testing Done:
CI passed on travis: https://travis-ci.org/pantsbuild/pants/builds/142920098

Jenkins seemed to be having issues.

Reviewed at https://rbcommons.com/s/twitter/r/4053/