Support for optional product requirements.
That is, you can now request products for the purpose of scheduling
in the round manager, but it won't be an error if nothing produces them.

This is then used in ide_gen.py, to request source products ('java',
'python', etc.) This provides an alternative solution to the problem
of failing on missing products when no codegen produces those source
products.  The previous solution (cb78b0ef7bc) was to have source roots
implicitly provide those products.  However this caused performance
issues because we have to scan the filesystem for source roots. This
solution does not have that problem.