[1.1.x] Clean up export deprecation warning on 1.1.x branch
**NOTE**: this is a patch to be applied to `1.1.x` branch.

This is so that the following warning is not shown because it's supposed to be pants internal code deprecation
for pants dev only, not for end users.
```
DeprecationWarning: DEPRECATED: jvm_distributions is deprecated in favor of preferred_jvm_distributions. will be removed in version 1.1.1.
```

Because on master this is already cleaned up by https://rbcommons.com/s/twitter/r/4084 along with a few
other things for `1.2.0-dev0` release, we can't cleanly cherry-pick, therefore making this a separate patch just for `1.1.x`.

Technically since this deprecation is supposed to happen `1.1.1`, and here it is `1.1.0`, we
do not need to clean up (adding Mateo in case he might have special reasons to keep it past 1.1.0)

Testing Done:
https://travis-ci.org/peiyuwang/pants/builds/146764000

Reviewed at https://rbcommons.com/s/twitter/r/4110/