Allow specification of an alternate javac location.
Currently we always use the javac embedded in the JDK
we're using to run zinc.

This change supports experimenting with unreleased compiler
versions without having to also build and run an unreleased,
and possibly unstable, runtime, which would be a much riskier
prospect.

Specifically, this allows running the (relatively stable but still
unreleased) javac 9 compiler on a Java 8 runtime, something that
Bazel currently supports for various reasons.

- Also remove a comment that was obsoleted long ago,
  in b15a7cbb to be precise.

- Also fix a couple of references to a non-public member of a
  base class.