A custom version of com.sun.tools.javac.api.JavacTool.
This class is what Zinc uses as a java compiler.

If we want to use some custom version of javac other
than the one embedded in the JDK we use to run Zinc, then
we just need to put that other javac on the classpath, before
any others, and let its JavacTool get classloaded first.

A future change will implement this, but to test that change
we need some alternative JavacTool that makes it easy to know
that it was invoked.  This change introduces such a JavacTool.

The future tests can verify that the custom JavacTool was indeed
invoked, by checking that the compilation attempt fails with
the message "Pants caused Zinc to load a custom JavacTool"
on stderr.

The reason this is in a separate change than the tests that use it
is that we need to publish it for those tests to use it, and that
must be done from master, so we can update the live pushdb.