Move the custom javac option to the Java subsystem.
It was initially an option on ZincCompile.  But putting it on Java
allows us to have plugins depend on the custom javac instead of
always depending on the JDK's tools.jar.  This solves mismatch
issues where we depend on the regular tools.jar when compiling
a plugin, but then that plugin runs against the custom javac.