Inject an automatic dep on junit for all junit_tests targets.
It's silly to require an explicit dep.

This required creating a new "JUnit" subsystem.  While there,
I moved the junit runner tool to that subsystem, so the runner
and the library are in the same place.

It's still clunkier than it needs to be to inject deps onto targets,
but that's a problem for another time.