JVM Compile bug fix
 https://rbcommons.com/s/twitter/r/3635/ introduced a regression where classpath entries for targets other than the target being compiled could end up on the compile classpath.

This fixes it by replacing the usages of the closed over iteration variable with the passed ctx variable.

It adds a regression test.

I also added a test that checks the contents of the classpath when --use-classpath-jars is passed, because I couldn't find a test that checked the used classpath for the jars.