Make util.objects.datatype classes not iterable
Datatypes are typically used as structs not collections. This makes it tricky to assume iterating over them is a reasonable operation. In fact, often it is not.

This patch makes datatype classes not iterable. It does so by overriding `__iter__` and the methods from namedtuple's template that expect `self` to be iterable.

Additional adjustments:

- ensure `super.__eq__` value is propagated even if it is NotImplemented
- add is_iterable=True to FileContent, since it is used as an iterable

Testing Done:
Added tests that checked the behavior, then iterated on failures from the initial CI run. Current CI away in PR.

Bugs closed: 3790

Reviewed at https://rbcommons.com/s/twitter/r/4163/

closes #3790