Deprecate the `subsystem_instance` utility function.
It was implemented as a contextmanager that resets subsystem
state in __exit__.  However this interacts badly with tests that
use that state outside the context. Several tests have to do
weird balancing acts to get around this.

Instead we introduce a global_subsystem_instance() that simply
inits and returns the global instance of a subsystem. The test
must rely on BaseTest.tearDown to reset state, or it can do so
explicitly itself if it must.

This change allows us to remove various contextmanagers sprinkled
around our tests, that existed only because subsystem_instance
was a contextmanager.  It also allows simplification of several
tests.

This is part 2 of my effort to simplify and standardize how we
create subsystems in tests.  See for part 1:

44be4da8bda29e21061e550b6529a346934b9b0f