Refactor memoization of the global distribution locator.
Previously it was memoized on the DistributionLocator class.
This meant that it survived resetting the subsystem instances
in tests, and required a special, non-obvious workaround.

Now we memoize it on the subsystem instance, so it gets cleaned
up when the subsystem is reset.  This change therefore also gets
rid of the workaround, which makes some tests simpler.

This change also adds subsystem initialization to a test that
needed it (unclear how that test worked before, possibly it
implicitly relied on some unclean state from a previous test).

It also removes the argument to _create_locator(), since it was
never used in practice.