Restore ChangedTargetGoalsIntegrationTest.
All of its tests were marked xfail but are now re-enabled with changes
to ensure the targets and git repo under test are isolated from the
host pants repo.

In order to facilitate integration tests using an ephemeral git repo,
a change is made to `build-support/pants_venv` to not rely on git to
find the host pants repo root directory.